feat(#S11P21D109-332): price type Long으로 변경

+++ android/domain/src/main/java/com/semonemo/domain/request/SellNftRequest.kt
@@ -1,9 +1,7 @@
 package com.semonemo.domain.request
 
-import java.math.BigInteger
-
 data class SellNftRequest(
     val nftId: Long,
-    val price: BigInteger,
+    val price: Long,
     val txHash: String,
 )
